% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_col_plants.R
\name{load_col_plants}
\alias{load_col_plants}
\title{Load Catalog of Life taxonomic data subsetted to vascular plants}
\usage{
load_col_plants(col_file)
}
\arguments{
\item{col_file}{Path to Catalog of Life taxonomic data
subsetted to vascular plants (downloaded from Catalog
of Life website)}
}
\value{
Tibble. taxonID, acceptedNameUsageID, parentNameUsageID
columns will be numeric, everything else character.
}
\description{
Load Catalog of Life taxonomic data subsetted to vascular plants
}
\examples{
\dontrun{
# Create a temporary working directory
library(fs)
temp_dir <- tempdir()
dir_create(temp_dir)

# Download Catalog of Life plants taxonomy
download.file(
  "http://www.catalogueoflife.org/DCA_Export/zip/archive-kingdom-plantae-phylum-tracheophyta-bl3.zip",
  fs::path(temp_dir, "archive-kingdom-plantae-phylum-tracheophyta-bl3.zip")
  )

# Unzip the downloaded folder
unzip(
  fs::path(temp_dir, "archive-kingdom-plantae-phylum-tracheophyta-bl3.zip"),
  files = "taxa.txt",
  exdir = temp_dir
)

# Load the CoL plants dataset
load_col_plants(fs::path(temp_dir, "taxa.txt"))

# Cleanup
fs::dir_delete(temp_dir)
}
}
